refactor(http_request): streamline HTTP request handling
- Renamed `HttpExecutor` to `Executor` for consistency.
- Unified `Response` handling by moving common logic to a single `Response` class to avoid redundancy.
- Simplified imports and file structure for clearer organization.
- Updated tests and references to reflect changes in class and module names.
