+++ api/core/workflow/nodes/http_request/__init__.py
@@ -1,4 +1,4 @@
 from .entities import BodyData, HttpRequestNodeAuthorization, HttpRequestNodeBody, HttpRequestNodeData
-from .http_request_node import HttpRequestNode
+from .node import HttpRequestNode
 
 __all__ = ["HttpRequestNodeData", "HttpRequestNodeAuthorization", "HttpRequestNodeBody", "BodyData", "HttpRequestNode"]
