+++ api/core/workflow/nodes/http_request/entities.py
@@ -1,10 +1,19 @@
 from collections.abc import Sequence
 from typing import Literal, Optional
 
+import httpx
 from pydantic import BaseModel, Field, ValidationInfo, field_validator
 
 from configs import dify_config
 from core.workflow.entities.base_node_data_entities import BaseNodeData
+
+NON_FILE_CONTENT_TYPES = (
+    "application/json",
+    "application/xml",
+    "text/html",
+    "text/plain",
+    "application/x-www-form-urlencoded",
+)
 
 
 class HttpRequestNodeAuthorizationConfig(BaseModel):
@@ -62,3 +71,51 @@ class HttpRequestNodeData(BaseNodeData):
     params: str
     body: Optional[HttpRequestNodeBody] = None
     timeout: Optional[HttpRequestNodeTimeout] = None
+
+
+class Response:
+    headers: dict[str, str]
+    response: httpx.Response
+
+    def __init__(self, response: httpx.Response):
+        self.response = response
+        self.headers = dict(response.headers)
+
+    @property
+    def is_file(self):
+        content_type = self.content_type
+        content_disposition = self.response.headers.get("Content-Disposition", "")
+
+        return "attachment" in content_disposition or (
+            not any(non_file in content_type for non_file in NON_FILE_CONTENT_TYPES)
+            and any(file_type in content_type for file_type in ("application/", "image/", "audio/", "video/"))
+        )
+
+    @property
+    def content_type(self) -> str:
+        return self.headers.get("Content-Type", "")
+
+    @property
+    def text(self) -> str:
+        return self.response.text
+
+    @property
+    def content(self) -> bytes:
+        return self.response.content
+
+    @property
+    def status_code(self) -> int:
+        return self.response.status_code
+
+    @property
+    def size(self) -> int:
+        return len(self.content)
+
+    @property
+    def readable_size(self) -> str:
+        if self.size < 1024:
+            return f"{self.size} bytes"
+        elif self.size < 1024 * 1024:
+            return f"{(self.size / 1024):.2f} KB"
+        else:
+            return f"{(self.size / 1024 / 1024):.2f} MB"
