+++ api/core/workflow/nodes/http_request/executor.py
@@ -11,10 +11,12 @@ from configs import dify_config
 from core.file import file_manager
 from core.helper import ssrf_proxy
 from core.workflow.entities.variable_pool import VariablePool
-from core.workflow.nodes.http_request.entities import (
+
+from .entities import (
     HttpRequestNodeAuthorization,
     HttpRequestNodeData,
     HttpRequestNodeTimeout,
+    Response,
 )
 
 BODY_TYPE_TO_CONTENT_TYPE = {
@@ -23,64 +25,9 @@ BODY_TYPE_TO_CONTENT_TYPE = {
     "form-data": "multipart/form-data",
     "raw-text": "text/plain",
 }
-NON_FILE_CONTENT_TYPES = (
-    "application/json",
-    "application/xml",
-    "text/html",
-    "text/plain",
-    "application/x-www-form-urlencoded",
-)
 
 
-class HttpExecutorResponse:
-    headers: dict[str, str]
-    response: httpx.Response
-
-    def __init__(self, response: httpx.Response):
-        self.response = response
-        self.headers = dict(response.headers)
-
-    @property
-    def is_file(self):
-        content_type = self.content_type
-        content_disposition = self.response.headers.get("Content-Disposition", "")
-
-        return "attachment" in content_disposition or (
-            not any(non_file in content_type for non_file in NON_FILE_CONTENT_TYPES)
-            and any(file_type in content_type for file_type in ("application/", "image/", "audio/", "video/"))
-        )
-
-    @property
-    def content_type(self) -> str:
-        return self.headers.get("Content-Type", "")
-
-    @property
-    def text(self) -> str:
-        return self.response.text
-
-    @property
-    def content(self) -> bytes:
-        return self.response.content
-
-    @property
-    def status_code(self) -> int:
-        return self.response.status_code
-
-    @property
-    def size(self) -> int:
-        return len(self.content)
-
-    @property
-    def readable_size(self) -> str:
-        if self.size < 1024:
-            return f"{self.size} bytes"
-        elif self.size < 1024 * 1024:
-            return f"{(self.size / 1024):.2f} KB"
-        else:
-            return f"{(self.size / 1024 / 1024):.2f} MB"
-
-
-class HttpExecutor:
+class Executor:
     method: Literal["get", "head", "post", "put", "delete", "patch"]
     url: str
     params: Mapping[str, str] | None
@@ -221,8 +168,8 @@ class HttpExecutor:
 
         return headers
 
-    def _validate_and_parse_response(self, response: httpx.Response) -> HttpExecutorResponse:
-        executor_response = HttpExecutorResponse(response)
+    def _validate_and_parse_response(self, response: httpx.Response) -> Response:
+        executor_response = Response(response)
 
         threshold_size = (
             dify_config.HTTP_REQUEST_NODE_MAX_BINARY_SIZE
@@ -260,7 +207,7 @@ class HttpExecutor:
         response = getattr(ssrf_proxy, self.method)(**request_args)
         return response
 
-    def invoke(self) -> HttpExecutorResponse:
+    def invoke(self) -> Response:
         # assemble headers
         headers = self._assembling_headers()
         # do http request
